% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddm.R
\name{drift_diffusion}
\alias{drift_diffusion}
\title{Generate a discrete-time diffusion process}
\usage{
drift_diffusion(
  bias = 0.5,
  driftrate = 1,
  decision_boundary = 2,
  ndt = 0.5,
  diffvar = 0.1,
  dt = 0.001,
  max_time = 6,
  ndt_noise = FALSE
)
}
\arguments{
\item{bias}{Starting point [0, 1]. 0.5 means no bias.}

\item{driftrate}{Average rate of evidence accumulation.}

\item{decision_boundary}{Symmetric absorbing boundaries.}

\item{ndt}{Non-decision time, in seconds.}

\item{diffvar}{Diffusion variance---must be > 0.}

\item{max_time}{Maximum duration of trial.}
}
\value{
A tibble containing \code{time}, the decision variable \code{dv} and the step number \code{steps}.
}
\description{
Generate a discrete-time diffusion process
}
\examples{
drift_diffusion(bias = 0.3, driftrate = 0.8)
drift_diffusion(bias = 0.5, driftrate = -1.0, decision_boundary = 2.5, diffvar = 0.8)

}
