% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_treatment.R
\name{simulate_treament}
\alias{simulate_treament}
\title{Generate a multilevel data for a treatment factor with two levels}
\usage{
simulate_treament(
  a = 3.5,
  b = -1,
  sigma_a = 1,
  sigma_b = 0.8,
  rho = -0.7,
  n_subjects = 10,
  n_trials = 20,
  sigma = 0.5
)
}
\arguments{
\item{a}{Average pre-treatment score}

\item{b}{Average difference between pre and post}

\item{sigma_a}{Std dev in intercepts}

\item{sigma_b}{Std dev in slopes -> set to 0.5}

\item{rho}{Correlation between intercepts and slopes}

\item{n_subjects}{Number of subjects}

\item{n_trials}{Number of trials per subjetc per treatment condition}

\item{sigma}{Std dev within subjects}
}
\value{
A tibble \code{d} containing the parameters, the linear predictor and the response for \code{n_trials} repetitions.
}
\description{
Generate a multilevel data for a treatment factor with two levels
}
\examples{
d <- simulate_data()
d_linpred <- d |>
    d |>group_by(subject, treatment) |> distinct(mu, .keep_all = TRUE)

}
